Update QIRO implementations MaxClique, MaxIndepSet
