+++ src/qrisp/algorithms/qiro/qiroproblems/qiroMaxClique.py
@@ -1,0 +1,150 @@
+"""
+\********************************************************************************
+* Copyright (c) 2023 the Qrisp authors
+*
+* This program and the accompanying materials are made available under the
+* terms of the Eclipse Public License 2.0 which is available at
+* http://www.eclipse.org/legal/epl-2.0.
+*
+* This Source Code may also be made available under the following Secondary
+* Licenses when the conditions for such availability set forth in the Eclipse
+* Public License, v. 2.0 are satisfied: GNU General Public License, version 2
+* with the GNU Classpath Exception which is
+* available at https://www.gnu.org/software/classpath/license.html.
+*
+* SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+********************************************************************************/
+"""
+
+from qrisp import rz, rzz, x
+import numpy as np
+import copy
+import networkx as nx
+from qrisp.algorithms.qiro.qiroproblems.qiro_utils import * 
+
+
+def create_maxClique_replacement_routine(res, graph, solutions, exclusions):
+    """
+    Creates a replacement routine for the problem structure, i.e., defines the replacement rules. 
+    See the `original paper <https://journals.aps.org/prxquantum/abstract/10.1103/PRXQuantum.5.020327>`_ for a description of the update rules.
+
+    Parameters
+    ----------
+    res : dict
+        Result dictionary of initial QAOA optimization procedure.
+    graph : nx.Graph
+        The graph defining the problem instance.
+    solutions : list
+        Qubits which were found to be positively correlated, i.e., part of the problem solution.
+    exclusions : list
+        Qubits which were found to be negatively correlated, i.e., not part of the problem solution, or contradict solution qubits in accordance with the update rules.  
+
+    Returns
+    -------
+    newgraph : nx.Graph
+        Updated graph for the problem instance.
+    solutions : list
+        Updated set of solutions to the problem.
+    sign : int
+        The sign of the correlation.
+    exclusions : list
+        Updated set of exclusions for the problem.
+        
+    """
+
+    orig_edges = [list(item) for item in graph.edges()]
+    orig_nodes = list(graph.nodes())
+
+    #get the max_edge and eval the sum and sign
+    max_item, sign = find_max(orig_nodes, orig_edges , res, solutions)
+
+    newgraph = copy.deepcopy(graph)
+
+    # we just directly remove vertices from the graph 
+    if isinstance(max_item, int):
+        if sign > 0:
+            border = list(graph.adj[max_item].keys())
+            border.append(max_item)
+            to_remove = [int(item) for item in graph.nodes() if item not in border]
+            newgraph.remove_nodes_from( [item for item in graph.nodes() if item not in border])
+            solutions.append(max_item)
+            exclusions += to_remove
+
+        elif sign < 0:
+            #remove item
+            newgraph.remove_node(max_item)
+            exclusions.append(max_item)
+
+    else:
+        if sign > 0:
+            #keep the two items in solution and remove all that are not adjacent to both
+            
+            intersect = list(set(list(graph.adj[max_item[0]].keys())) & set(list(graph.adj[max_item[0]].keys())))
+            intersect.append(max_item[0])
+            intersect.append(max_item[1])
+            to_remove = [int(item) for item in graph.nodes() if item not in intersect]
+            newgraph.remove_nodes_from([item for item in graph.nodes() if item not in intersect])
+            solutions.append(max_item[0])
+            solutions.append(max_item[1])
+            exclusions += to_remove
+
+        elif sign < 0:
+            #remove all that do not border on either! node
+            union = list(graph.adj[max_item[0]].keys())
+            union += list(graph.adj[max_item[1]].keys())
+            union.append(max_item[0])
+            union.append(max_item[1])
+            to_remove = [int(item) for item in graph.nodes() if item not in union]
+            #to_delete = [item for item in graph.nodes() if item not in union]
+            newgraph.remove_nodes_from([item for item in graph.nodes() if item not in union])
+            exclusions += to_remove
+
+    return newgraph, solutions, sign, exclusions
+
+
+def create_maxClique_cost_operator_reduced(graph, solutions=[]):
+    r"""
+    Creates the ``cost_operator`` for the problem instance.
+    This operator is adjusted to consider qubits that were found to be a part of the problem solution.
+
+    Parameters
+    ----------
+    G : nx.Graph
+        The graph for the problem instance.
+    solutions : list
+        Qubits which were found to be positively correlated, i.e., part of the problem solution.
+    
+    Returns
+    -------
+    cost_operator : function
+        A function receiving a :ref:`QuantumVariable` and a real parameter $\gamma$. This function performs the application of the cost operator.
+
+    """
+    G_compl = nx.complement(graph)
+    def cost_operator(qv, gamma):
+        for pair in list(G_compl.edges()):
+            rzz(3*gamma, qv[pair[0]], qv[pair[1]])
+            rz(-gamma, qv[pair[0]])
+            rz(-gamma, qv[pair[1]])
+        for i in graph.nodes():
+            if not i in solutions:
+                rz(gamma, qv[i])
+
+    return cost_operator
+
+
+
+
+
+
+
+
+
+"""
+BM: 18 nodes, seed 99
+0.5933101570400762
+normal
+0.08909071112167886
+
+
+"""
